% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanKesselR-package.R
\docType{package}
\name{vanKesselR-package}
\alias{vanKesselR}
\alias{vanKesselR-package}
\title{vanKesselR: Streamline Common Data Tasks to Increase Productivity}
\description{
The package provides lightweight utilities to increase productivity by streamlining common, repetitive tasks.
}
\author{
\strong{Maintainer}: Hugo van Kessel \email{hugovankessel@gmail.com} (\href{https://orcid.org/0000-0001-9794-6385}{ORCID})

}
\keyword{internal}
